+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/token_controller.ex
@@ -4,7 +4,7 @@ defmodule BlockScoutWeb.API.V2.TokenController do
   alias BlockScoutWeb.AccessHelper
   alias BlockScoutWeb.API.V2.{AddressView, TransactionView}
   alias Explorer.{Chain, Repo}
-  alias Explorer.Chain.{Address, Token, Token.Instance}
+  alias Explorer.Chain.{Address, BridgedToken, Token, Token.Instance}
   alias Indexer.Fetcher.TokenTotalSupplyOnDemand
 
   import BlockScoutWeb.Chain,
@@ -18,7 +18,12 @@ defmodule BlockScoutWeb.API.V2.TokenController do
     ]
 
   import BlockScoutWeb.PagingHelper,
-    only: [delete_parameters_from_next_page_params: 1, token_transfers_types_options: 1, tokens_sorting: 1]
+    only: [
+      chain_ids_filter_options: 1,
+      delete_parameters_from_next_page_params: 1,
+      token_transfers_types_options: 1,
+      tokens_sorting: 1
+    ]
 
   import Explorer.MicroserviceInterfaces.BENS, only: [maybe_preload_ens: 1]
 
@@ -32,6 +37,27 @@ defmodule BlockScoutWeb.API.V2.TokenController do
          {:not_found, {:ok, token}} <- {:not_found, Chain.token_from_address_hash(address_hash, @api_true)} do
       TokenTotalSupplyOnDemand.trigger_fetch(address_hash)
 
+      conn
+      |> token_response(token, address_hash)
+    end
+  end
+
+  if Application.compile_env(:explorer, Explorer.Chain.BridgedToken)[:enabled] do
+    defp token_response(conn, token, address_hash) do
+      if token.bridged do
+        bridged_token = Repo.get_by(BridgedToken, home_token_contract_address_hash: address_hash)
+
+        conn
+        |> put_status(200)
+        |> render(:bridged_token, %{token: {token, bridged_token}})
+      else
+        conn
+        |> put_status(200)
+        |> render(:token, %{token: token})
+      end
+    end
+  else
+    defp token_response(conn, token, _address_hash) do
       conn
       |> put_status(200)
       |> render(:token, %{token: token})
@@ -279,6 +305,25 @@ defmodule BlockScoutWeb.API.V2.TokenController do
     conn
     |> put_status(200)
     |> render(:tokens, %{tokens: tokens, next_page_params: next_page_params})
+  end
+
+  def bridged_tokens_list(conn, params) do
+    filter = params["q"]
+
+    options =
+      params
+      |> paging_options()
+      |> Keyword.merge(chain_ids_filter_options(params))
+      |> Keyword.merge(tokens_sorting(params))
+      |> Keyword.merge(@api_true)
+
+    {tokens, next_page} = filter |> BridgedToken.list_top_bridged_tokens(options) |> split_list_by_page()
+
+    next_page_params = next_page |> next_page_params(tokens, delete_parameters_from_next_page_params(params))
+
+    conn
+    |> put_status(200)
+    |> render(:bridged_tokens, %{tokens: tokens, next_page_params: next_page_params})
   end
 
   defp put_owner(token_instances, holder_address),